+++ .htaccess
@@ -1,4 +1,4 @@
-DirectoryIndex index.php 
+DirectoryIndex index.php
 
 RewriteEngine On
-RewriteRule ^(fr|en)?/?([a-zA-Z0-9_\-]*)?/?$ index.php?lang=$1&page=$1&%{QUERY_STRING}
+RewriteRule ^([a-zA-Z0-9_\-]*)?/?$ index.php?page=$1&%{QUERY_STRING}
