+++ _classes/Articles.php
@@ -6,9 +6,9 @@ class Articles
     public $sentence;
     public $content;
     public $date;
-    public $admin;
-    public $category;
-    public $image;
+    public $id_admin;
+    public $id_category;
+    public $id_image;
 
     function __construct($id)
     {
@@ -16,7 +16,11 @@ class Articles
 
         $id = str_secur($id);
 
-        $reqArticle = $db->prepare('SELECT `articles`.*, `images`.`id_article`, `admins`.`id`, `categories`.* FROM `articles` LEFT JOIN `images` ON `images`.`id_article` = `articles`.`id` LEFT JOIN `admins` ON `admins`.`id_image` = `images`.`id` LEFT JOIN `categories` ON `articles`.`id_category` = `categories`.`id`
+        $reqArticle = $db->prepare('SELECT `articles`.*, `images`.`image`, `categories`.`nom`, `admins`.`firstname`
+        FROM `articles` 
+            LEFT JOIN `images` ON `images`.`id_article` = `articles`.`id` 
+            LEFT JOIN `categories` ON `articles`.`id_category` = `categories`.`id` 
+            LEFT JOIN `admins` ON `admins`.`id_image` = `images`.`id`
         WHERE `articles`.`id` = ?
         
         ');
@@ -25,11 +29,14 @@ class Articles
         $data = $reqArticle->fetch();
         $this->id = $id;
         $this->title = $data['title'];
-        $this->content = $data['content'];
         $this->sentence = $data['sentence'];
+        $this->content = $data['content'];
         $this->date = $data['date'];
-        $this->admin = $data['firstname'] . '' . $data['lastname'];
-        $this->category = $data['id_category'];
+        $this->id_admin = $data['id_admin'];
+        $this->id_category = $data['id_category'];
+        $this->id_image = $data['id_image'];
+
+
         
     }
 
@@ -44,7 +51,11 @@ class Articles
 
             global $db;
 
-            $reqArticle = $db->prepare('SELECT `articles`.*, `images`.`id_article`, `admins`.`id`, `categories`.* FROM `articles` LEFT JOIN `images` ON `images`.`id_article` = `articles`.`id` LEFT JOIN `admins` ON `admins`.`id_image` = `images`.`id` LEFT JOIN `categories` ON `articles`.`id_category` = `categories`.`id`
+            $reqArticle = $db->prepare('SELECT `articles`.*, `images`.`image`, `categories`.`nom`, `admins`.`firstname`
+            FROM `articles` 
+                LEFT JOIN `images` ON `images`.`id_article` = `articles`.`id` 
+                LEFT JOIN `categories` ON `articles`.`id_category` = `categories`.`id` 
+                LEFT JOIN `admins` ON `admins`.`id_image` = `images`.`id`;`
             ');
             $reqArticle->execute([]);
             return $reqArticle->fetchAll();
