--- _lang/en/footer.json
@@ -1,3 +1,0 @@
-{
-    "footer" : "Footer"
-}
