--- _lang/en/header.json
@@ -1,3 +1,0 @@
-{
-    "header" : "Header"
-}
