--- _lang/en/home.json
@@ -1,3 +1,0 @@
-{
-    "home" : "Home"
-}
