--- _lang/fr/footer.json
@@ -1,3 +1,0 @@
-{
-    "footer" : "Pied de page"
-}
