--- _lang/fr/header.json
@@ -1,3 +1,0 @@
-{
-    "header" : "en-tête"
-}
