--- _lang/fr/home.json
@@ -1,3 +1,0 @@
-{
-    "home" : "Accueil"
-}
