--- controllers/admin_controllers.php
@@ -1,8 +1,0 @@
-<?php
-
-$shop = Shop::getAllShop();
-$allArticle = Articles::getAllArticles();
-$allCategory = Categories::getAllCategories();
-$allAdmin  = Admin::getAllAdmin();
-$allImage = Images::getAllImages();
-$allAuthors = Authors::getAllAuthors();
