--- controllers/home_controller.php
@@ -1,6 +1,0 @@
-<?php 
-
-
-
-
-
