--- models/home_model.php
@@ -1,10 +1,0 @@
-<?php
-
-//  !TODO ici on y mettra les classe propre a la page
-/**
- * [Description home]
- */
-class home {
-
-
-}
