+++ views/includes/header.php
@@ -23,6 +23,8 @@
                 <li>
                     <a href="index.php?page=contact">Contact</a>
                 </li>
+
+                
               
             </ul>
         </nav>
