--- views/test_grid.view.php
@@ -1,38 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <?php include_once 'views/includes/head.php'?>
-
-    <title>Document</title>
-</head>
-
-<body>
-    <h1>hello</h1>
-    <div class="containerXL">
-        <div class="container">
-
-            <div class="row">
-                <div class="col-6">col-6</div>
-                <div class="col-6">col-6</div>
-            </div>
-
-            <div class="row">
-                <div class="col-8">col-8</div>
-                <div class="col-2">col-2</div>
-                <div class="col-2">col-2</div>
-            </div>
-
-            <div class="row">
-                <div class="col-4">col-4</div>
-                <div class="col-4">col-4</div>
-                <div class="col-2">col-2</div>
-                <div class="col-2">col-2</div>
-            </div>
-        </div>
-    </div>
-</body>
-
-</html>
